+++ cmd/argoexec/commands/root.go
@@ -1,17 +1,9 @@
 package commands
 
 import (
-	"bufio"
-	"bytes"
-	"encoding/json"
-	"fmt"
-	"io"
 	"os"
-	"strings"
 
 	"github.com/argoproj/argo"
-	"github.com/argoproj/argo/errors"
-	wfv1 "github.com/argoproj/argo/pkg/apis/workflow/v1alpha1"
 	"github.com/argoproj/argo/util/cmd"
 	"github.com/argoproj/argo/workflow/common"
 	"github.com/argoproj/argo/workflow/executor"
@@ -69,14 +61,6 @@ func initExecutor() *executor.WorkflowExecutor {
 		podAnnotationsPath = GlobalArgs.podAnnotationsPath
 	}
 
-	var wfTemplate wfv1.Template
-
-	// Read template
-	err := getTemplateFromPodAnnotations(podAnnotationsPath, &wfTemplate)
-	if err != nil {
-		log.Fatalf("Error getting template %v", err)
-	}
-
 	config, err := getClientConfig(GlobalArgs.kubeConfig)
 	if err != nil {
 		panic(err.Error())
@@ -97,95 +81,16 @@ func initExecutor() *executor.WorkflowExecutor {
 
 	// Initialize workflow executor
 	wfExecutor := executor.WorkflowExecutor{
-		PodName:   podName,
-		Template:  wfTemplate,
-		ClientSet: clientset,
-		Namespace: namespace,
+		PodName:            podName,
+		ClientSet:          clientset,
+		Namespace:          namespace,
+		PodAnnotationsPath: podAnnotationsPath,
+	}
+	err = wfExecutor.LoadTemplate()
+	if err != nil {
+		panic(err.Error())
 	}
 	yamlBytes, _ := yaml.Marshal(&wfExecutor.Template)
 	log.Infof("Executor (version: %s) initialized with template:\n%s", argo.GetVersion(), string(yamlBytes))
 	return &wfExecutor
-}
-
-// Open the Kubernetes downward api file and
-// read the pod annotation file that contains template and then
-// unmarshal the template
-func getTemplateFromPodAnnotations(annotationsPath string, template *wfv1.Template) error {
-	// Read the annotation file
-	file, err := os.Open(annotationsPath)
-	if err != nil {
-		fmt.Printf("ERROR opening annotation file from %s\n", annotationsPath)
-		return errors.InternalWrapError(err)
-	}
-
-	defer func() {
-		_ = file.Close()
-	}()
-	reader := bufio.NewReader(file)
-
-	// Prefix of template property in the annotation file
-	prefix := fmt.Sprintf("%s=", common.AnnotationKeyTemplate)
-
-	for {
-		// Read line-by-line
-		var buffer bytes.Buffer
-
-		var l []byte
-		var isPrefix bool
-		for {
-			l, isPrefix, err = reader.ReadLine()
-			buffer.Write(l)
-
-			// If we've reached the end of the line, stop reading.
-			if !isPrefix {
-				break
-			}
-
-			// If we're just at the EOF, break
-			if err != nil {
-				break
-			}
-		}
-
-		// The end of the annotation file
-		if err == io.EOF {
-			break
-		}
-
-		line := buffer.String()
-
-		// Read template property
-		if strings.HasPrefix(line, prefix) {
-			// Trim the prefix
-			templateContent := strings.TrimPrefix(line, prefix)
-
-			// This part is a bit tricky in terms of unmarshalling
-			// The content in the file will be something like,
-			// `"{\"type\":\"container\",\"inputs\":{},\"outputs\":{}}"`
-			// which is required to unmarshal twice
-
-			// First unmarshal to a string without escaping characters
-			var templateString string
-			err = json.Unmarshal([]byte(templateContent), &templateString)
-			if err != nil {
-				fmt.Printf("Error unmarshalling annotation into template string, %s, %v\n", templateContent, err)
-				return errors.InternalWrapError(err)
-			}
-
-			// Second unmarshal to a template
-			err = json.Unmarshal([]byte(templateString), template)
-			if err != nil {
-				fmt.Printf("Error unmarshalling annotation into template, %s, %v\n", templateString, err)
-				return errors.InternalWrapError(err)
-			}
-			return nil
-		}
-	}
-
-	if err != io.EOF {
-		return errors.InternalWrapError(err)
-	}
-
-	// If reaching here, then no template prefix in the file
-	return errors.InternalErrorf("No template property found from annotation file: %s", annotationsPath)
 }
